split everything into files
