+++ lib/actions/splash.rb
@@ -1,0 +1,9 @@
+class Splash < Action
+  def use(attacker, targets)
+    puts action_msg
+    targets.each do |target|
+      damage_result = calculate_damage(attacker, self.dmg_mod, target)
+      puts "#{damage_result}"
+    end
+  end
+end
